Fixed file path for testing and classes
