+++ src/css/styles.scss
@@ -1,0 +1,208 @@
+@import 'base';
+
+// --- deskTop ---
+
+// --- header ---
+body {
+  background-color: $primary-color;
+}
+
+header {
+  position: absolute;
+  top: 0;
+  display: flex;
+  background-color: $secondary-color;
+  width: 100%;
+  height: 15vh;
+}
+
+.user-title {
+  width: 30%;
+  display: flex;
+  justify-content: space-around;
+  align-items: center;
+}
+
+.page-title {
+  font-size: 1.5em;
+  margin-left: 4%;
+  width: 70%;
+  display: flex;
+  justify-content: flex-start;
+}
+ // --- login ---
+
+.login-card {
+  margin-top: 12%;
+  @include card(space-around, row, 50vh, 40%);
+  background-color: $card-color;
+  box-shadow: 2px 10px 10px 2px $cardShadow-color;
+  align-items: center;
+  justify-content: center;
+}
+
+.styling-box {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.login-form {
+  width: 100%;
+  height: 100%;
+  flex-direction: column;
+  display: flex;
+  justify-content: space-around;
+  align-items: center;
+}
+
+.login-button {
+  font-size: 1em;
+  box-shadow: 2px 5px 5px 2px $cardShadow-color;
+  background-color: $secondary-color;
+  @include button(2% 15% 2% 15%, 0);
+}
+
+//  --- main/cards
+
+.user-dashboard {
+  width: 100%;
+  margin-top: 14%;
+  display: flex;
+  justify-content: space-around;
+}
+
+.card {
+  @include card(space-around, column, 50vh, 40%);
+  background-color: $card-color;
+  box-shadow: 2px 10px 10px 2px $cardShadow-color;
+  margin: 10%, 5%, 5%, 5%;
+}
+
+
+.search-room {
+  align-items: center;
+}
+
+.user-history {
+  text-align: center;
+}
+
+.date-input {
+  width: 100%;
+}
+
+.booking-submit {
+  font-size: 1em;
+  box-shadow: 2px 5px 5px 2px $cardShadow-color;
+  background-color: $secondary-color;
+  @include button(2% 15% 2% 15%, 0);
+}
+
+//  ---- bottom styling ----
+
+.footer {
+  position: fixed;
+  bottom: 0;
+  background-color: $secondary-color;
+  width: 100%;
+  height: 12vh;
+}
+// tablet
+@media screen and (max-width: 925px) {
+
+  .page-title {
+    font-size: 1.25em;
+    margin-left: 2%;
+  }
+
+}
+
+
+// ---- phone ----
+@media screen and (max-width: 414px) {
+
+  header {
+    position: relative;
+    top: 0;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    background-color: $secondary-color;
+    width: 100%;
+    height: 10vh;
+  }
+
+  h1 {
+    margin: 0;
+    padding: 0;
+  }
+
+  h2 {
+    margin: 0;
+    padding: 0;
+  }
+
+  .page-title {
+    font-size: 1em;
+    justify-content: center;
+    align-items: center;
+    padding: 0;
+  }
+
+  .page-title {
+    font-size: .75em;
+    justify-content: center;
+    text-align: center;
+    margin: 0;
+    padding: 0;
+  }
+
+  .user-dashboard{
+    margin-top: 0%;
+    flex-direction: column;
+  }
+
+  .card {
+    @include card(space-around, column, 75vh, 100%);
+    background-color: $card-color;
+    box-shadow: 2px 10px 10px 2px $cardShadow-color;
+    margin: 18%, 2%, 20%, 2%;
+  }
+
+  .user-history {
+    margin: 15% 0 30% 0;
+  }
+
+  .login-card {
+    margin-top: 5%;
+    @include card(space-around, row, 75vh, 100%);
+    background-color: $card-color;
+    box-shadow: 2px 10px 10px 2px $cardShadow-color;
+    align-items: center;
+    justify-content: center;
+  }
+
+  .styling-box {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+  }
+
+  .login-form {
+    width: 100%;
+    height: 100%;
+    flex-direction: column;
+    display: flex;
+    justify-content: space-around;
+    align-items: center;
+  }
+}
+
+
+//  ---- utility stuff ----
+
+.hidden {
+  display: none;
+}
