+++ test/booking-Test-Data.js
@@ -1,29 +1,3 @@
-const customersTestData = [
-  {"id":1, "name": "Person One"},
-  {"id":2, "name": "Person Two"},
-  {"id":3, "name": "Person Three"},
-  {"id":4, "name": "Person Four"},
-  {"id":5, "name": "Person Five"},
-  {"id":6, "name": "Person Six"},
-  {"id":7, "name": "Person Seven"},
-  {"id":8, "name": "Person Eight"},
-  {"id":9, "name": "Person Nine"},
-  {"id":10, "name": "Person Ten"}
-];
-
-const roomsTestData = [
-  {"number":1,"roomType":"residential suite","bidet":true,"bedSize":"queen","numBeds":1,"costPerNight":350.50},
-  {"number":2,"roomType":"suite","bidet":false,"bedSize":"full","numBeds":2,"costPerNight":475.50},
-  {"number":3,"roomType":"single room","bidet":false,"bedSize":"king","numBeds":1,"costPerNight":490.75},
-  {"number":4,"roomType":"single room","bidet":false,"bedSize":"queen","numBeds":1,"costPerNight":430.25},
-  {"number":5,"roomType":"single room","bidet":true,"bedSize":"queen","numBeds":2,"costPerNight":340.15}
-  {"number":6,"roomType":"junior suite","bidet":true,"bedSize":"queen","numBeds":1,"costPerNight":400.00},
-  {"number":7,"roomType":"junior suite","bidet":false,"bedSize":"king","numBeds":2,"costPerNight":495.50},
-  {"number":8,"roomType":"single room","bidet":false,"bedSize":"queen","numBeds":1,"costPerNight":375.55},
-  {"number":9,"roomType":"residential suite","bidet":false,"bedSize":"queen","numBeds":1,"costPerNight":340.00},
-  {"number":10,"roomType":"single room","bidet":false,"bedSize":"full","numBeds":2,"costPerNight":430.99}
-];
-
 const bookingsTestData = [
   {"id":"5fwrgu4i7k55hl6t5","userID":1,"date":"2020/01/24","roomNumber":1,"roomServiceCharges":[]},
   {"id":"5fwrgu4i7k55hl6t6","userID":1,"date":"2020/01/10","roomNumber":2,"roomServiceCharges":[]},
@@ -57,7 +31,4 @@ const bookingsTestData = [
   {"id":"5fwrgu4i7k55hl6tf","userID":10,"date":"2020/01/25","roomNumber":10,"roomServiceCharges":[]}
 ];
 
-
 export default bookingsTestData;
-export default roomsTestData;
-export default customersTestData;
