+++ test/booking-test.js
@@ -1,8 +1,14 @@
 import chai from 'chai';
 const expect = chai.expect;
-import bookingsTestData from './testing-data';
-import Booking from './Booking';
+import bookingsTestData from './booking-Test-Data';
+import Booking from '../src/Booking';
 
 describe('Booking', function() {
+
+  it.skip('Should be a function', function() {
+    expect(Booking).to.be.a('function');
+  });
+
+
 
 });
