+++ test/bookingRepo-test.js
@@ -1,8 +1,8 @@
 import chai from 'chai';
 const expect = chai.expect;
-import bookingsTestData from './testing-data';
-import Booking from './Booking';
-import BookingRepo from './BookingRepo';
+import bookingsTestData from './booking-Test-Data';
+import Booking from '../src/Booking';
+import BookingRepo from '../src/BookingRepo';
 
 describe('BookingRepo', function() {
