+++ test/manager-test.js
@@ -1,6 +1,6 @@
 import chai from 'chai';
 const expect = chai.expect;
-import Manager from './Manager';
+import Manager from '../src/Manager';
 
 describe('Manager', function() {
