+++ test/room-test-data.js
@@ -1,0 +1,14 @@
+const roomsTestData = [
+  {"number":1,"roomType":"residential suite","bidet":true,"bedSize":"queen","numBeds":1,"costPerNight":358.4},
+  {"number":2,"roomType":"suite","bidet":false,"bedSize":"full","numBeds":2,"costPerNight":477.38},
+  {"number":3,"roomType":"single room","bidet":false,"bedSize":"king","numBeds":1,"costPerNight":491.14},
+  {"number":4,"roomType":"single room","bidet":false,"bedSize":"full","numBeds":1,"costPerNight":429.44},
+  {"number":5,"roomType":"single room","bidet":true,"bedSize":"twin","numBeds":2,"costPerNight":340.17},
+  {"number":6,"roomType":"junior suite","bidet":true,"bedSize":"queen","numBeds":1,"costPerNight":397.02},
+  {"number":7,"roomType":"single room","bidet":false,"bedSize":"queen","numBeds":2,"costPerNight":231.46},
+  {"number":8,"roomType":"junior suite","bidet":false,"bedSize":"king","numBeds":1,"costPerNight":261.26},
+  {"number":9,"roomType":"single room","bidet":true,"bedSize":"queen","numBeds":1,"costPerNight":200.39},
+  {"number":10,"roomType":"suite","bidet":false,"bedSize":"twin","numBeds":1,"costPerNight":497.64}
+];
+
+export default roomsTestData;
