+++ test/room-test.js
@@ -1,7 +1,7 @@
 import chai from 'chai';
 const expect = chai.expect;
-import roomsTestData from './testing-data';
-import Room from './Room';
+import roomsTestData from './room-test-data';
+import Room from '../src/Room';
 
 describe('Room', function() {
