+++ test/roomRepo-test.js
@@ -1,8 +1,8 @@
 import chai from 'chai';
 const expect = chai.expect;
-import roomsTestData from './testing-data';
-import Room from './Room';
-import RoomRepo from './RoomRepo';
+import roomsTestData from './room-test-data';
+import Room from '../src/Room';
+import RoomRepo from '../src/RoomRepo';
 
 describe('RoomRepo', function() {
