+++ test/user-Test-Data.js
@@ -1,0 +1,14 @@
+const customersTestData = [
+  {"id":1, "name": "Person One"},
+  {"id":2, "name": "Person Two"},
+  {"id":3, "name": "Person Three"},
+  {"id":4, "name": "Person Four"},
+  {"id":5, "name": "Person Five"},
+  {"id":6, "name": "Person Six"},
+  {"id":7, "name": "Person Seven"},
+  {"id":8, "name": "Person Eight"},
+  {"id":9, "name": "Person Nine"},
+  {"id":10, "name": "Person Ten"}
+];
+
+export default customersTestData;
