+++ test/user-test.js
@@ -1,7 +1,7 @@
 import chai from 'chai';
 const expect = chai.expect;
-import customersTestData  from './testing-data';
-import User from './User';
+import customersTestData  from './user-Test-Data';
+import User from '../src/User';
 
 describe('User', function() {
